Fix three bugs found by the new unit tests
- parseIsoTime ignored the timezone offset in the input string, converting
  the local part with a fixed UTC offset. Harmless with GitHub's Z-suffixed
  release dates, but incorrect for any non-zero offset.
- getNextOrNull returned the first element of the list instead of null when
  the predicate matched nothing (indexOfFirst returns -1). All current call
  sites in the marker code cannot hit the no-match branch, so no behavior
  changes in the editor.
- Double.roundToDecimalDigit converted through Float, losing precision when
  the scaled value exceeds 2^24 (e.g. property values written to label files
  for positions beyond ~2.8 minutes with 2 decimal digits).

The tests that previously pinned the buggy behavior now assert the corrected
behavior, with a new large-value case for the Double rounding.

+++ src/jvmMain/kotlin/com/sdercolin/vlabeler/util/Math.kt
@@ -8,5 +8,5 @@ fun Float.roundToDecimalDigit(n: Int?): Float = if (n == null) this else {
 }
 
 fun Double.roundToDecimalDigit(n: Int?): Double = if (n == null) this else {
-    10.0.pow(n).let { times(it).roundToLong().toFloat().div(it) }
+    10.0.pow(n).let { times(it).roundToLong().toDouble().div(it) }
 }
